Isolate handlebars helpers to own util
