+++ module/bunkers-and-badasses.mjs
@@ -7,7 +7,7 @@ import { BNBItemSheet } from "./sheets/item-sheet.mjs";
 // Import helper/utility classes and constants.
 import { preloadHandlebarsTemplates } from "./helpers/templates.mjs";
 import { BNB } from "./helpers/config.mjs";
-import { Dropdown } from "./helpers/dropdown.mjs";
+import { HandlebarsHelperUtil } from "./helpers/handlebarsHelperUtil.mjs"
 
 
 /* -------------------------------------------- */
@@ -116,104 +116,7 @@ Hooks.once('init', async function() {
 /* -------------------------------------------- */
 /*  Handlebars Helpers                          */
 /* -------------------------------------------- */
-
-// If you need to add Handlebars helpers, here are a few useful examples:
-Handlebars.registerHelper('concat', function() {
-  let outStr = '';
-  for (let arg in arguments) {
-    if (typeof arguments[arg] != 'object') {
-      outStr += arguments[arg];
-    }
-  }
-  return outStr;
-});
-
-Handlebars.registerHelper('capitalize', function(str) {
-  return str.charAt(0).toUpperCase() + str.slice(1);
-});
-
-Handlebars.registerHelper('toLowerCase', function(str) {
-  return str.toLowerCase();
-});
-
-Handlebars.registerHelper('toUpperCase', function(str) {
-  return str.toUpperCase();
-});
-
-Handlebars.registerHelper('enrich', function(str) {
-  return TextEditor.enrichHTML(str, {async: false});;
-});
-
-Handlebars.registerHelper('toArray', (...values) => {
-  // Omit the Handlebars options object.
-  return values.slice(0, values.length - 1);
-});
-
-Handlebars.registerHelper('toFavoredElementObject', function(...values) {
-  return {label: values[0], favored: values[1]};
-});
-
-
-Handlebars.registerHelper('hpTitle', function(str) {
-  if (str === "flesh")
-    str = "health";
-  return str.charAt(0).toUpperCase() + str.slice(1);
-});
-
-Handlebars.registerHelper('hpToRecoveryTitle', function(str, doCapitalize) {
-  if (str === "flesh")
-    str = "regen";
-  else if (str === "shield")
-    str = "recharge";
-  else if (str === "armor")
-    str = "repair";
-  else if (str === "bone")
-    str = "regrow";
-  else if (str === "eridian")
-    str = "reinvigorate";
-
-  if (doCapitalize)
-    return str.charAt(0).toUpperCase() + str.slice(1);
-  else
-    return str;
-});
-
-Handlebars.registerHelper('getBestHealthShade', function(str) {
-  if (str === "flesh")
-    str = "light";
-  else if (str === "shield")
-    str = "light";
-  else if (str === "armor")
-    str = "light";
-  else if (str === "bone")
-    str = "dark";
-  else if (str === "eridian")
-    str = "light";
-  return str;
-});
-
-Handlebars.registerHelper('shortName', function(str) {
-  let check = str.toLowerCase();
-  if (check === 'submachine gun')
-    return 'SMG';
-  else if (check === 'combat rifle')
-    return 'Rifle';
-  else if (check === 'sniper rifle')
-    return 'Sniper';
-  else if (check === 'rocket launcher')
-    return 'RL';
-  else
-    return str;
-});
-
-// Dropdown related helpers.
-Handlebars.registerHelper('dropdownComponentClass', function(componentType) {
-  return Dropdown.getComponentClass(componentType);
-});
-Handlebars.registerHelper('dropdownComponentCss', function(componentType) {
-  return Dropdown.getComponentCss(componentType);
-});
-
+HandlebarsHelperUtil.prepareHandlebarsHelpers();
 
 /* -------------------------------------------- */
 /*  Ready Hook                                  */
